+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/BrowserToolbar.kt
@@ -25,6 +25,7 @@ import mozilla.components.browser.toolbar.edit.EditToolbar
 import mozilla.components.concept.toolbar.AutocompleteDelegate
 import mozilla.components.concept.toolbar.AutocompleteResult
 import mozilla.components.concept.toolbar.Toolbar
+import mozilla.components.concept.toolbar.Toolbar.PermissionHighlights
 import mozilla.components.support.base.android.Padding
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.ui.autocomplete.AutocompleteView
@@ -110,6 +111,14 @@ class BrowserToolbar @JvmOverloads constructor(
     override var siteSecure: Toolbar.SiteSecurity
         get() = display.siteSecurity
         set(value) { display.siteSecurity = value }
+
+    override var permissionHighlights: PermissionHighlights = PermissionHighlights.NONE
+        set(value) {
+            if (field != value) {
+                display.setPermissionIndicator(value)
+                field = value
+            }
+        }
 
     override var siteTrackingProtection: Toolbar.SiteTrackingProtection =
         Toolbar.SiteTrackingProtection.OFF_GLOBALLY