+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/display/TrackingProtectionIconView.kt
@@ -126,7 +126,7 @@ internal class TrackingProtectionIconView @JvmOverloads constructor(
     }
 }
 
-private class Update(
+internal class Update(
     val drawable: Drawable?,
     @StringRes val contentDescription: Int?,
     val visible: Boolean